Add version correlation/comparison tables.
One of the more obscure uses of update server logs is to examine if there is a correlation between using an old version of WordPress/PHP and updating plugins. For example, you can take an old WordPress version like 4.1 and look at which version(s) of your plugin are most common on those sites. In my case, it turns out that 90% of people who are using WP 4.1 are also using an outdated version of my plugin (i.e. it was released more than a year ago). This suggests that we could raise minimum WP version requirements without inconveniencing existing users because those users don't install updates anyway.