Ignore up to 30 consecutive malformed lines.
Previously, when the parser encountered a bad line, it would throw an exception and stop execution. This was a problem when trying to parse logs that contain malformed requests (e.g. due to someone trying to hack the server; yes, that has happened). Most of the file was still good, but the malformed lines caused the parser to stop half-way.

Instead, lets ignore small amounts of parser errors and keep going. However, if there are lots of consecutive errors, that probably means the file is not actually a valid wp-update-server log, so lets stop then.